ajout prototype:authentification et gestion des roles par spatie

+++ prototype/app/Http/Controllers/ArticleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Article;
+use App\Models\Category; 
 use App\Services\ArticleService;
 
 
@@ -22,7 +23,57 @@ class ArticleController extends Controller
         return view('articles.index', $data);
     }
 
+    public function create()
+    {
+       $categories = Category::all(); 
+       return view('articles.create', compact('categories'));
+    }
+
+    public function store(Request $request)
+  {
+    if(!auth()->user()->can('create articles')){
+        abort(403,  "Vous n'avez pas la permission de créer un article.");
+    }
+
+    // Validation
+      $validated = $request->validate([
+        'title' => 'required|string|max:180',
+        'slug' => 'nullable|string|max:200|unique:articles,slug',
+        'excrept' => 'nullable|string|max:500',
+        'content' => 'nullable|string',
+        'status' => 'required|in:publié,brouillon',
+        'categories' => 'required|array',
+      ]);
+
+       // Génération du slug si vide
+        $slug = $validated['slug'] ?? \Str::slug($validated['title']);
+
+      // Création de l'article
+      $article = Article::create([
+        'user_id' => auth()->id(),
+        'title' => $validated['title'],
+        'slug' => $slug,
+        'excrept' => $validated['excrept'] ?? '',
+        'content' => $validated['content'] ?? '',
+        'status' => $validated['status'],
+      ]);
+
+       // Lier les catégories (sync au lieu d'attach)
+       $article->categories()->sync($validated['categories']);
+
+       return redirect()->route('articles.index')->with('success', 'Article ajouté avec succès.');
+    }
+
+
     public function destroy(Article $article){
+        if(!auth()->user()->can('delete articles')){
+            abort(403, 'Vous n\'avez pas la permission de supprimer cet article.');
+        }
+        
+        if(auth()->user()->hasRole('auteur') && auth()->id() !== $article->user_id){
+            abort(403, 'vous ne peouvez supprimer que vos propores articles.');
+        }
+        
         $this->articleService->delete($article);
         return redirect()->route('articles.index')->with('success',' Article supprimé avec succés.');
     }
